fix(retirement-tooling): split recommender_retirement_packet to clear the god-files gate
CI's sentrux-capability-gate / diagnosis.hospital flagged the single-file
E02 port as a new god file (1031 loc, rule: loc>800). Split along the PS1
originals' own script boundaries instead of arbitrarily: this file keeps
generate()/verify() (New-/Test-RecommenderRetirementPacket.ps1),
recommender_retirement_restore.rs gets restore_legacy_branch()
(Restore-RecommenderLegacyBranch.ps1), recommender_retirement_diff.rs gets
the delete-hunk/result-text computation, and recommender_retirement_shared.rs
holds the constants and find_bounded_block primitive all three need.

Verification:
- code-intel sentrux --operation check --repo .: 'All rules passed',
  'God files: 33 -> 32', 'No degradation detected' -- confirms the actual
  gate that failed in CI now passes locally.
- cargo test -p code-intel --bin code-intel recommender_retirement: 9
  passed (correctly redistributed across the 4 files), 1 ignored
- Re-ran the ignored end-to-end integration test: PASSED
- cargo test -p code-intel --bin code-intel (full sweep): 972/972 passed
- cargo fmt -p code-intel -- --check: clean
- code-intel lint hardcoded-paths: OK (292 files)
