+++ crates/code-intel-cli/src/main.rs
@@ -61,7 +61,10 @@ mod ponytail_gate;
 mod project_context;
 mod project_orientation_benchmark;
 mod providers;
+mod recommender_retirement_diff;
 mod recommender_retirement_packet;
+mod recommender_retirement_restore;
+mod recommender_retirement_shared;
 mod repin;
 mod repowise_hooks;
 mod repowise_i18n_proxy;
